Create new utils file for logger
